api_Ingest/requirements.txt
functions-framework==3.*
google-cloud-storage==2.16.0
requests==2.31.0
pandas==1.4.3
openpyxl==3.0.10